Use f-strings and dataclasses
